NXP-9260 Integrate locking into CMIS CheckIn/CheckOut/CancelCheckOut operations

+++ nuxeo-opencmis-impl/src/main/java/org/nuxeo/ecm/core/opencmis/impl/server/NuxeoCmisService.java
@@ -1704,6 +1704,7 @@ public class NuxeoCmisService extends AbstractCmisService {
         try {
             coreSession.saveDocument(doc);
             DocumentRef ver = doc.checkIn(option, checkinComment);
+            doc.removeLock();
             coreSession.save();
             objectIdHolder.setValue(getIdFromDocumentRef(ver));
         } catch (ClientException e) {
@@ -1733,6 +1734,7 @@ public class NuxeoCmisService extends AbstractCmisService {
         try {
             coreSession.saveDocument(doc);
             DocumentRef ver = doc.checkIn(option, checkinComment);
+            doc.removeLock();
             coreSession.save();
             return getIdFromDocumentRef(ver);
         } catch (ClientException e) {
@@ -1778,6 +1780,11 @@ public class NuxeoCmisService extends AbstractCmisService {
                 throw new CmisConstraintException("Already checked out: "
                         + objectId);
             }
+            if (pwc.isLocked()) {
+                throw new CmisInvalidArgumentException(
+                        "Cannot check out since currently locked: " + objectId);
+            }
+            pwc.setLock();
             pwc.checkOut();
             coreSession.save();
             return pwc.getId();
@@ -1808,6 +1815,7 @@ public class NuxeoCmisService extends AbstractCmisService {
             } else {
                 // restore and keep checked in
                 coreSession.restoreToVersion(docRef, verRef, true, true);
+                doc.removeLock();
             }
             coreSession.save();
         } catch (ClientException e) {
@@ -1981,7 +1989,21 @@ public class NuxeoCmisService extends AbstractCmisService {
     @Override
     public void deleteObjectOrCancelCheckOut(String repositoryId,
             String objectId, Boolean allVersions, ExtensionsData extension) {
-        deleteObject(repositoryId, objectId, allVersions, extension);
+        try {
+            DocumentModel doc = getDocumentModel(objectId);
+            DocumentRef docRef = doc.getRef();
+            // find last version
+            DocumentRef verRef = coreSession.getLastDocumentVersionRef(docRef);
+            // If doc has versions, is locked, and is checkedOut, then it was likely
+            // explicitly checkedOut so invoke cancelCheckOut not delete
+            if (verRef != null && doc.isLocked() && doc.isCheckedOut()) {
+                cancelCheckOut(repositoryId, objectId, extension);
+            } else {
+                deleteObject(repositoryId, objectId, allVersions, extension);
+            }
+        } catch (ClientException e) {
+            throw new CmisRuntimeException(e.toString(), e);
+        }
     }
 
 }
